refactor: signup styledInput
